Criação de rotinas de cadastro e atualização de dados e recuperação de senha.

+++ app/src/main/java/br/com/smartpromos/api/general/SmartRepo.java
@@ -100,7 +100,7 @@ public interface SmartRepo {
 
     @FormUrlEncoded
     @POST("/prc/prc.php?ctl=mobile&acao=getLocalizacao")
-    void getLocalizacao(@Field("customer") int customer,
+    void getLocalizacao(@Field("customer") long customer,
                    Callback<LocalizacaoResponse> localizacaoResponse);
 
 }
