+++ app/src/main/java/br/com/smartpromos/api/general/request/CupomRequest.java
@@ -8,7 +8,7 @@ import com.google.gson.annotations.SerializedName;
 public class CupomRequest {
 
     @SerializedName("id")
-    private int id_coupon;
+    private long id_coupon;
 
     @SerializedName("nome")
     private String name;
@@ -35,12 +35,12 @@ public class CupomRequest {
     private int quantity;
 
     @SerializedName("sale")
-    private int sale;
+    private long sale;
 
     @SerializedName("msg")
     private MensagemRequest mensagem;
 
-    public CupomRequest(int id_coupon, String name, String description, String confirmationmessage, String path_img, int type, int notification, String mensage, int quantity, int sale, MensagemRequest mensagem) {
+    public CupomRequest(long id_coupon, String name, String description, String confirmationmessage, String path_img, int type, int notification, String mensage, int quantity, long sale, MensagemRequest mensagem) {
         this.id_coupon = id_coupon;
         this.name = name;
         this.description = description;
@@ -63,7 +63,7 @@ public class CupomRequest {
     }
 
 
-    public int getId_coupon() {
+    public long getId_coupon() {
         return id_coupon;
     }
 
@@ -99,11 +99,11 @@ public class CupomRequest {
         return quantity;
     }
 
-    public int getSale() {
+    public long getSale() {
         return sale;
     }
 
-    public void setId_coupon(int id_coupon) {
+    public void setId_coupon(long id_coupon) {
         this.id_coupon = id_coupon;
     }
 
@@ -139,7 +139,7 @@ public class CupomRequest {
         this.quantity = quantity;
     }
 
-    public void setSale(int sale) {
+    public void setSale(long sale) {
         this.sale = sale;
     }
 }
