+++ app/src/main/java/br/com/smartpromos/api/general/response/ClienteResponse.java
@@ -10,7 +10,7 @@ import java.io.Serializable;
 public class ClienteResponse {
 
     @SerializedName("cpf")
-    private int doc_id;
+    private long doc_id;
 
     @SerializedName("nome")
     private String first_name;
@@ -49,7 +49,7 @@ public class ClienteResponse {
     private MensagemResponse mensagem;
 
 
-    public ClienteResponse(int doc_id, String first_name, String last_name, int gender, int birthday, int birthday_month, int birthday_yaer, int sale_radius, int get_offers, int stay_logged_in, String email, String phone, MensagemResponse mensagem) {
+    public ClienteResponse(long doc_id, String first_name, String last_name, int gender, int birthday, int birthday_month, int birthday_yaer, int sale_radius, int get_offers, int stay_logged_in, String email, String phone, MensagemResponse mensagem) {
         this.doc_id = doc_id;
         this.first_name = first_name;
         this.last_name = last_name;
@@ -74,7 +74,7 @@ public class ClienteResponse {
     }
 
 
-    public int getDoc_id() {
+    public long getDoc_id() {
         return doc_id;
     }
 
@@ -122,7 +122,7 @@ public class ClienteResponse {
         return phone;
     }
 
-    public void setDoc_id(int doc_id) {
+    public void setDoc_id(long doc_id) {
         this.doc_id = doc_id;
     }
