+++ app/src/main/java/br/com/smartpromos/api/general/response/LocalizacaoResponse.java
@@ -10,7 +10,7 @@ import com.google.gson.annotations.SerializedName;
 public class LocalizacaoResponse {
 
     @SerializedName("id")
-    private int id_locale;
+    private long id_locale;
 
     @SerializedName("bairro")
     private String neighborwood;
@@ -40,12 +40,12 @@ public class LocalizacaoResponse {
     private String addresstype;
 
     @SerializedName("customer")
-    private int customer;
+    private long customer;
 
     @SerializedName("msg")
     private MensagemResponse mensagem;
 
-    public LocalizacaoResponse(int id_locale, String neighborwood, String city, String state, int zip_code, String country, String address, String number_address, int type, String addresstype, int customer, MensagemResponse mensagem) {
+    public LocalizacaoResponse(long id_locale, String neighborwood, String city, String state, int zip_code, String country, String address, String number_address, int type, String addresstype, long customer, MensagemResponse mensagem) {
         this.id_locale = id_locale;
         this.neighborwood = neighborwood;
         this.city = city;
@@ -60,11 +60,11 @@ public class LocalizacaoResponse {
         this.mensagem = mensagem;
     }
 
-    public int getId_locale() {
+    public long getId_locale() {
         return id_locale;
     }
 
-    public void setId_locale(int id_locale) {
+    public void setId_locale(long id_locale) {
         this.id_locale = id_locale;
     }
 
@@ -140,11 +140,11 @@ public class LocalizacaoResponse {
         this.addresstype = addresstype;
     }
 
-    public int getCustomer() {
+    public long getCustomer() {
         return customer;
     }
 
-    public void setCustomer(int customer) {
+    public void setCustomer(long customer) {
         this.customer = customer;
     }
